Unbelievable rooky mistake. Only run cron once per day
It was trying to collect the data and send mailouts once per
minute for the full hour between 12am and 1pm. So, what would
happen is there would be multiple rake tasks running at the
same time until all the memory ran out and the server crashed.
Oh what fun!