Add empty files directory to openaustralia role
Needed so that generate-certificates.sh can drop certificates here.

Since it's empty, we need to add a file. Since we want to make sure we
never commit the pem files, we may as well make the file a .gitignore
and use it to make sure the pem files don't get committed.