Find all the things that differ across environments and variablise.
I've chosen to make heavy use of the ternary pattern here in order to
have just one "main config" task. This avoids having to declare a
matching list of variables in two places.

Later, I'd like to come back and rework this into a Role that we can
call multiple times, passing in a stage each time; this will mean we
remove the with_items from the task and simplify the ternary calls, but
the list of variables will remain unchanged.