Add new API key for use on website
This adds the new key required for use with
openaustralia/planningalerts/1327

The existing key is renamed to be the "email" key, as we can't change
the key that's been used on emails that have already been sent out.

The existing key name is applied to the new key, so that new web
traffic will start using the new key immediately.