Remove managed db on linode
This was a bad idea. Very dissapointed with what the managed db is on linode. No ability to see what is going on. How much memory is it using? No idea. What are the database logs saying? No idea. Can i give the server some more memory please. No way! You need to rebuild a whole new cluster and migrate it. You have got to be kidding.