Remove the plausible role as it's not used
Ansible was attempting to install this role on vagrant VMs that don't need it
because there's no Plausible host group defined in the Vagrantfile. As we're not
currently using it, the cleanest solution seems to be to clean up the unused
code.

In the future, if we do move ahead with this, we can pull the work that was done
on this out of this commit.